==> 03_Notebooks/02_Desarrollo/01-Set up.ipynb <==
### Importacion paquetes

import pandas as pd
import numpy as np
import os

%config IPCompleter.greedy = True

### Crear el directorio del proyecto.

raiz = r'C:\Users\manug\Desktop\Data Science\DataSets\Wall-Mart\\'

nombre_dir = 'wallmart'

path = raiz + nombre_dir

 try:
     os.mkdir(path)
     os.mkdir(path + '/01_Documentos')
     os.mkdir(path + '/02_Datos')
     os.mkdir(path + '/02_Datos/01_Originales')
     os.mkdir(path + '/02_Datos/02_Validacion')
     os.mkdir(path + '/02_Datos/03_Trabajo')
     os.mkdir(path + '/02_Datos/04_Caches')
     os.mkdir(path + '/03_Notebooks')
     os.mkdir(path + '/03_Notebooks/01_Funciones')
     os.mkdir(path + '/03_Notebooks/02_Desarrollo')
     os.mkdir(path + '/03_Notebooks/03_Sistema')
     os.mkdir(path + '/04_Modelos')
     os.mkdir(path + '/05_Resultados')
     os.mkdir(path + '/09_Otros')
    
 except OSError:
     print ("La creación del directorio %s ha fallado" % path)
 else:
     print ("Se ha creado satisfactoriamente el directorio %s " % path)

os.chdir(path)

### Cargar los datos iniciales

nombre_fichero_datos = 'features.csv'
ruta_completa = path + '/02_Datos/01_Originales/' + nombre_fichero_datos

features = pd.read_csv(ruta_completa, parse_dates = ['Date'])
features

nombre_fichero_datos = 'stores.csv'
ruta_completa = path + '/02_Datos/01_Originales/' + nombre_fichero_datos

stores = pd.read_csv(ruta_completa)
stores.head()

nombre_fichero_datos = 'train.csv'
ruta_completa = path + '/02_Datos/01_Originales/' + nombre_fichero_datos

train = pd.read_csv(ruta_completa)
train

nombre_fichero_datos = 'test.csv'
ruta_completa = path + '/02_Datos/01_Originales/' + nombre_fichero_datos

test = pd.read_csv(ruta_completa)
test

### Se unen los dataframes features y stores

df = features.merge(right = stores, how = 'left', on = 'Store').reset_index(drop = True).set_index('Date')
df.info()

### Se unen los dataframes df con train (para obtener la var Dept y la target)

import pandas as pd

# Asegurarnos de que la columna 'Date' en el DataFrame 'train' esté en formato de fecha
train['Date'] = pd.to_datetime(train['Date'])

# Convertir el índice de 'df' a una columna para facilitar el merge
df.reset_index(inplace=True)

# Realizar la combinación de los DataFrames utilizando 'merge'
df_combined = pd.merge(df, train[['Store', 'Dept', 'Date', 'Weekly_Sales']], 
                       on=['Store', 'Date'], how='left')



df = df_combined.copy()

df.set_index('Date', inplace = True)

### Extraer y reservar el dataset de validación

val = df.sort_index().loc['2012-07-26':'2013-07-26']
val

nombre_fichero_validacion = 'validacion.csv'
ruta_completa = path + '/02_Datos/02_Validacion/' + nombre_fichero_validacion

val.to_csv(ruta_completa)

### Extraer y  guardarel dataset de trabajo

trabajo = df.drop(val.index)
trabajo

nombre_fichero_trabajo = 'trabajo.csv'

ruta_completa = path + '/02_Datos/03_Trabajo/' + nombre_fichero_trabajo

trabajo.to_csv(ruta_completa)
