==> 03_Notebooks/02_Desarrollo/04-Transformacion de datos.ipynb <==
import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
%matplotlib inline

from sklearn.preprocessing import OneHotEncoder
from category_encoders import TargetEncoder

from sklearn.preprocessing import MinMaxScaler

%config IPCompleter.greedy = True

ruta_proyecto = r'C:\Users\manug\Desktop\Data Science\DataSets\Wall-Mart\wallmart\\'

nombre_cat = 'cat_resultado_eda.pickle'
nombre_num = 'num_resultado_eda.pickle'

cat = pd.read_pickle(ruta_proyecto + '02_Datos/03_Trabajo/' + nombre_cat)
num = pd.read_pickle(ruta_proyecto + '02_Datos/03_Trabajo/' + nombre_num)

df = pd.concat([cat,num], axis = 1)
df

## Creación de nuevas variables
* Lags y ventanas m

#### Creación lags

lags = [1, 2, 3, 4]
for lag in lags:
    df[f'weekly_sales_lag_{lag}'] = df.groupby(['store', 'dept'])['weekly_sales'].shift(lag)
    df[f'temperature_lag_{lag}'] = df.groupby(['store'])['temperature'].shift(lag)
    df[f'fuel_price_lag_{lag}'] = df.groupby(['store'])['fuel_price'].shift(lag)
    df[f'cpi_lag_{lag}'] = df.groupby(['store'])['cpi'].shift(lag)
    df[f'unemployment_lag_{lag}'] = df.groupby(['store'])['unemployment'].shift(lag)
    df[f'isholiday_lag_{lag}'] = df.groupby(['store'])['isholiday'].shift(lag)

#### Creación medias movil

windows = [2, 4, 8] 

for window in windows:
    df[f'weekly_sales_mm_{window}'] = df.groupby(['store', 'dept'])['weekly_sales'].transform(lambda x: x.rolling(window).mean())
    df[f'temperature_mm_{window}'] = df.groupby(['store'])['temperature'].transform(lambda x: x.rolling(window).mean())
    df[f'fuel_price_mm_{window}'] = df.groupby(['store'])['fuel_price'].transform(lambda x: x.rolling(window).mean())
    df[f'cpi_mm_{window}'] = df.groupby(['store'])['cpi'].transform(lambda x: x.rolling(window).mean())
    df[f'unemployment_mm_{window}'] = df.groupby(['store'])['unemployment'].transform(lambda x: x.rolling(window).mean())

df.info()

#### Eliminación de nuevos nulos generados

df.dropna(inplace=True)

df.info()

#### Identicicación de la target

target = df.weekly_sales

#### Divición cat y num

cat = df.select_dtypes(include = 'O')
num = df.select_dtypes(exclude = 'O')

## Transformación de categóricas

# One Hot Encoding

cat

var_ohe = ['isholiday', 'type', 'isholiday_lag_1', 'isholiday_lag_2',
       'isholiday_lag_3', 'isholiday_lag_4']

ohe = OneHotEncoder(sparse_output=False,handle_unknown='ignore')

cat_ohe = ohe.fit_transform(cat[var_ohe])

cat_ohe

cat_ohe = pd.DataFrame(cat_ohe, columns=ohe.get_feature_names_out())

cat_ohe

### Target Encoding

var_te = ['isholiday', 'type', 'isholiday_lag_1', 'isholiday_lag_2',
       'isholiday_lag_3', 'isholiday_lag_4']

te = TargetEncoder(min_samples_leaf=100, return_df=False)

pd.set_option('future.no_silent_downcasting', True)
cat_te = te.fit_transform(cat[var_te], y=target)

nombre_te = [variable + '_te' for variable in var_te]

cat_te = pd.DataFrame(cat_te, columns=nombre_te)

cat_te

### Reescalar Variables

#### Min-Max Scaler

#### Variables a reescalar

num.columns

target = df['weekly_sales']

store = df['store']

variables_incluir_mms = ['temperature', 'fuel_price', 'markdown1', 'markdown2', 'markdown3',
       'markdown4', 'markdown5', 'cpi', 'unemployment', 'size', 'dept',
       'weekly_sales_lag_1', 'temperature_lag_1',
       'fuel_price_lag_1', 'cpi_lag_1', 'unemployment_lag_1',
       'weekly_sales_lag_2', 'temperature_lag_2', 'fuel_price_lag_2',
       'cpi_lag_2', 'unemployment_lag_2', 'weekly_sales_lag_3',
       'temperature_lag_3', 'fuel_price_lag_3', 'cpi_lag_3',
       'unemployment_lag_3', 'weekly_sales_lag_4', 'temperature_lag_4',
       'fuel_price_lag_4', 'cpi_lag_4', 'unemployment_lag_4',
       'weekly_sales_mm_2', 'temperature_mm_2', 'fuel_price_mm_2', 'cpi_mm_2',
       'unemployment_mm_2', 'weekly_sales_mm_4', 'temperature_mm_4',
       'fuel_price_mm_4', 'cpi_mm_4', 'unemployment_mm_4', 'weekly_sales_mm_8',
       'temperature_mm_8', 'fuel_price_mm_8', 'cpi_mm_8', 'unemployment_mm_8']

mms = MinMaxScaler()

df_mms = mms.fit_transform(df[variables_incluir_mms])

nombres_mms = [variable + '_mms' for variable in variables_incluir_mms]
df_mms = pd.DataFrame(df_mms, columns=nombres_mms)
df_mms

## Se unifican los datasets transformados

df_tablon = pd.concat([cat_ohe, cat_te, df_mms, target.reset_index(), store.reset_index()], axis=1)

df_tablon

## Guardar df_tablon

ruta_df_tablon = ruta_proyecto + '02_Datos/03_Trabajo/' + 'df_tablon.pickle'

df_tablon.to_pickle(ruta_df_tablon)

df_tablon.store
