==> 03_Notebooks/02_Desarrollo/05-Preseleccion de variables.ipynb <==
import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
%matplotlib inline

from sklearn.feature_selection import mutual_info_regression
from sklearn.feature_selection import RFE
from xgboost import XGBRegressor
from sklearn.inspection import permutation_importance

%config IPCompleter.greedy = True

ruta_proyecto = r'C:\Users\manug\Desktop\Data Science\DataSets\Wall-Mart\wallmart\\'

nombre_df = 'df_tablon.pickle'

df = pd.read_pickle(ruta_proyecto + '02_Datos/03_Trabajo/' + nombre_df)

df.info()

### METODOS SUPERVISADOS

#### Preparación x e y

target = 'weekly_sales'

x = df.drop(columns = ['date', 'store', 'weekly_sales']).copy()
y = df[target].copy()

#### Mutual Information

mutual_selector = mutual_info_regression(x, y)

mutual_selector

def ranking_mi(mutual_selector, modo = 'tabla'):
    #Maqueta el ranking
    ranking_mi = pd.DataFrame(mutual_selector, index = x.columns).reset_index()
    ranking_mi.columns = ['variable','importancia_mi']
    ranking_mi = ranking_mi.sort_values(by = 'importancia_mi', ascending = False)
    ranking_mi['ranking_mi'] = np.arange(0,ranking_mi.shape[0])
    #Muestra la salida
    if modo == 'tabla':
        return(ranking_mi)
    else:
        g = ranking_mi[0:45].importancia_mi.sort_values().plot.barh(figsize = (12,8))
        g.set_yticklabels(ranking_mi[0:45].sort_values(by = 'importancia_mi').variable)
        plt.show()
        return(g)

ranking_mi = ranking_mi(mutual_selector, modo = 'tabla');

ranking_mi.tail(50)

#### Seleccionar las variables que pasan

posicion_variable_limite = 46

entran_mi = ranking_mi.iloc[0:posicion_variable_limite].variable

x_mi = x[entran_mi].copy()
x_mi

#### Se evaluaron los sigueintes metodos de preselaccion de variable:
* Mutual Information
* Recursive Feature Elimination
* Permutation Importance

Se seleccionó el metodo Mutual Information para la preseleccion de variables por lo que fueron los resultados arrojados de cada metodo

#### Seleccionar método final

x_preseleccionada = x_mi

#### Incorporar variables de segmentación

x_preseleccionada = pd.concat([df[['date', 'store', 'weekly_sales']], x_preseleccionada], axis = 1)

x_preseleccionada

#### Guardar datasets

nombre_x_preseleccionado = ruta_proyecto + '/02_Datos/03_Trabajo/' + 'x_preseleccionado.pickle'
nombre_y_preseleccionado = ruta_proyecto + '/02_Datos/03_Trabajo/' + 'y_preseleccionado.pickle'

x_preseleccionada.to_pickle(nombre_x_preseleccionado)

y_preseleccionado = y.copy()
y_preseleccionado.to_pickle(nombre_y_preseleccionado)
