==> 03_Notebooks/02_Desarrollo/06-Modelizacion.ipynb <==
### Importación paquetes

import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
%matplotlib inline

from sklearn.model_selection import TimeSeriesSplit

from sklearn.pipeline import Pipeline
from sklearn.linear_model import LinearRegression
from sklearn.ensemble import RandomForestRegressor
from xgboost import XGBRegressor
from sklearn.experimental import enable_hist_gradient_boosting  # for HistGradientBoostingRegressor
from sklearn.ensemble import HistGradientBoostingRegressor

from sklearn.model_selection import GridSearchCV


%config IPCompleter.greedy = True

ruta_proyecto = r'C:\Users\manug\Desktop\Data Science\DataSets\Wall-Mart\wallmart\\'

nombre_x = 'x_preseleccionado.pickle'
nombre_y = 'y_preseleccionado.pickle'

x = pd.read_pickle(ruta_proyecto + '02_Datos/03_Trabajo/' + nombre_x)
y = pd.read_pickle(ruta_proyecto + '02_Datos/03_Trabajo/' + nombre_y)

### Modelizar para una tienda

x

x = x.loc[x.store == 1]

x

y = y[y.index.isin(x.index)]

var_modelizar = x.columns.to_list()[2:]

### Definir la validación

time_cv = TimeSeriesSplit(3, test_size = 8)

### Crear el pipe y el diccionario de algorimos, parámetros y valores a testar

# Definición del Pipeline
pipe = Pipeline([('algoritmo', LinearRegression())])

# Definición de los hiperparámetros a evaluar
grid = [
    {'algoritmo': [LinearRegression()]},


    {'algoritmo': [HistGradientBoostingRegressor()],
     'algoritmo__learning_rate': [0.01, 0.025, 0.05, 0.1],
     'algoritmo__max_iter': [50, 100, 200],
     'algoritmo__max_depth': [5, 10, 20],
     'algoritmo__min_samples_leaf': [500],
     'algoritmo__l2_regularization': [0, 0.25, 0.5, 0.75, 1]}
]

grid_search = GridSearchCV(estimator= pipe, 
                            param_grid = grid, 
                            cv = time_cv, 
                            scoring = 'neg_mean_squared_error',
                            verbose = 0,
                            n_jobs = -1)

modelo = grid_search.fit(x[var_modelizar],y)

pd.DataFrame(grid_search.cv_results_).sort_values(by = 'rank_test_score')

resultados

x
